pyproject.toml
[tool.poetry]
name = "google.drive"
version = "0.1.0"
description = ""
authors = ["github <github>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.23.2"
PyDrive = "^1.3.1"
streamlit = "^1.12.0"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
